fixed error where any command but ##end and ##start would infinite loop
