+++ .vscode/settings.json
@@ -29,5 +29,6 @@
 		"editor.rulers": [
 			80
 		]
-	}
+	},
+	"editor.renderWhitespace": "all"
 }
