+++ src/stdin_data_init.c
@@ -67,22 +67,3 @@ void		append_data_node(t_data **data, char *line)
 		current->next = new;
 	}
 }
-
-// void	set_path_length(t_anthill *anthill)
-// {
-// 	// TODO naming
-// 	t_pathlist	*current_path;
-// 	t_path		*current;
-
-// 	current_path = anthill->paths;
-// 	while (current_path)
-// 	{
-// 		current = current_path->path;
-// 		while (current)
-// 		{
-// 			current_path->length++;
-// 			current = current->next;
-// 		}
-// 		current_path = current_path->next;
-// 	}
-// }
