+++ visualiser/makefile
@@ -6,9 +6,19 @@ INCLUDES = -I./includes -I../lib/libft
 RPAD = 45
 LIBRARY = -L./lib -lSDL2 -lSDL2_gfx -lSDL2_image -L../lib/libft -lft
 
+
+
 all: -visual
 
 -visual:
+
 	@printf "%-$(RPAD)s" "Compiling visual executable. "
 	@$(CC) $(FLAGS) $(SRC_FILES) $(INCLUDES) $(LIBRARY) ./main.c -o visual
 	$(DONE)
+
+	@printf "%-$(RPAD)s" "Updating dylib path in executable. "
+	@install_name_tool -change /usr/local/opt/sdl2_image/lib/libSDL2_image-2.0.0.dylib @executable_path/lib/libSDL2_image-2.0.0.dylib visual
+	@install_name_tool -change /usr/local/opt/sdl2/lib/libSDL2-2.0.0.dylib @executable_path/lib/libSDL2-2.0.0.dylib visual
+	@install_name_tool -change /usr/local/opt/sdl2/lib/libSDL2-2.0.0.dylib @executable_path/lib/libSDL2-2.0.0.dylib visual
+	$(DONE)
+
